refactor: extract the issue state lookup into a shared module
check:roadmap のGitHub API問い合わせ（リポジトリの解決・トークン・
タイムアウト・到達できない番号の扱い）を tools/issue-state.mjs に出した。
roadmap以外のIssue参照を見る次のチェックと共有するため。
